second major facelift, relatively production ready code
contributions by chatgpt helped make this fully functional and updated
my code to a more organized object oriented and docker compatible
version. The find_tdc001_ports method is far more powerful and useful
for our purposes. The next goal will be to control this externally using
a docker container to hopefully get everything we want operational as a
proof-of-concept. The main method was also updated to only operate if
this is run as a standalone script with also expanded functionality.