added presentation on major docker concepts (How To Docker #1)
How To Docker #1 is a presentation covering up until the problem arose
of docker desktop not allowing USB devices to work out of the box. In
the readme, I will include instructions to getting this existing main
branch running on windows as-is.

While there is no UI, and this hasn't been made into a fast API
interface, the code, just as it had run through normal python before,
does run exactly the same through the terminal with the basic controller
commands. Pretty fancy.