v1.8ui, fully usable, minus some features. Divided into multiple scripts.
The version 1.8 ui includes:

-seeking the backend docker container automatically
-seeking the devices connected to it, ability to select from a dropdown
-all possible things in version 1.1 including the basic movements and
functions in the controller

Due to complexity, it is now being divided into many scripts for
readability and reusability purposes. main_window.py,
run.py,constants.py, and the others are for this reason.

Also in the works are restore windows on boot, restoring settings for a
particular previously used device, warnings for if a device has shut off
or has been moved by someone else. To run main_window.py, just run
run.py