operational containerized frontend with novnc output
major updates:

-frontend has full capabilities of searching for backend and detecting
connected devices, movement and everything else works

-lost some quality of life features such as auto initialization, power
loss recovery popups, and memorizing previous settings on the move to
docker

current work:

-building the "windows" directory for windows users to boot the
application and prebuilt stable release images

-adding in a bat file for the frontend and backend initialization (you
can run them on separate computers), with the addition of a shortcut on
your desktop menu

-the run file will also automatically open the url in your preferred
browser to the novnc server for interaction with the gui.

-For the backend, a similar feature will instead redirect to the fastapi
/docs for the backend to provide an intuitive beginning to what is going
on with that script basically.