fixed typo in service in powershell script
the docker service name is not "gui", it is "tdc_frontend" which is
served in the docker-compose file. I just had to correct this one line
error. There will be a lot of these 1 line error corrections since I am
doing this from linux and uploading to my local repository on a remote
windows machine.