updated readme for the startup/shutdown scripts
made it a bit neater and more comprehensive as a guide for this
particular part with some general info on the scripts and issues ahead.
As of now, I haven't fixed the issues it cites, but hopefully someone in
the future can read through this project and understand everything I did
based on my github and these notes.