updated readme's for all OS directories (MacOS,Linux,Windows), launcher, shutdown, save, and load scripts v1.0 ready
Got all the platforms to be able to launch or build docker images, save
docker images as a zip file to the Docker-images directory, and load
images from that directory as well (with the option of renaming or
changing the tag of the image).

 The new readme files were rewritten to reflect all information for each
of the operating systems and their quirks.

 I cannot write documentation for usbipd for macos since I only have
catalina, and that supports an older version of docker (4.0.0). I need
like 4.35+ to be able to do passthrough, but that's okay for now.