allowed load and save scripts to be executable in linux, included an easter egg for past development
my first docker container I ever built, the legacy image, is included in
docker images. It isn't fancy, but it was a first basic intro to
interacting with the command line inside a docker container and
controlling some hardware. You can refer to the original readme.md for
the first stable version to see how this worked.